requirements.txt
cvxopt
image
ipython
ipythonblocks
ipywidgets
jupyter
keras
matplotlib
networkx
numpy
opencv-python
pandas
pillow
pytest-cov
qpsolvers
scipy
sortedcontainers
tensorflow